chore(website): add documentation comments to various functions in playground - fix some linting issues - small code reorganization
