+++ packages/website/src/components/lib/markdown.ts
@@ -1,7 +1,7 @@
-import { parseESLintRC } from '../config/utils';
 import type { ConfigModel } from '../types';
+import { parseESLintRC } from './parseConfig';
 
-export function createSummary(
+function createSummary(
   value: string,
   title: string,
   type: 'ts' | 'json',
@@ -14,41 +14,46 @@ export function createSummary(
   return code;
 }
 
-function createSummaryJson(
-  obj: ConfigModel['tsconfig'] | ConfigModel['eslintrc'],
-  field: string,
-  title: string,
-): string {
+function createSummaryJson(obj: string, field: string, title: string): string {
   if (obj && Object.keys(obj).length > 0) {
     return createSummary(obj, title, 'json', 10);
   }
   return '';
 }
 
-export function genVersions(state: ConfigModel): string {
+/**
+ * Generate a table of versions
+ */
+export function genVersions(tsVersion: string): string {
   return [
     '| package | version |',
     '| -- | -- |',
     `| \`@typescript-eslint/eslint-plugin\` | \`${process.env.TS_ESLINT_VERSION}\` |`,
     `| \`@typescript-eslint/parser\` | \`${process.env.TS_ESLINT_VERSION}\` |`,
-    `| \`TypeScript\` | \`${state.ts}\` |`,
+    `| \`TypeScript\` | \`${tsVersion}\` |`,
     `| \`ESLint\` | \`${process.env.ESLINT_VERSION}\` |`,
     `| \`node\` | \`web\` |`,
   ].join('\n');
 }
 
+/**
+ * Create a markdown string that user can copy and paste into an issue
+ */
 export function createMarkdown(state: ConfigModel): string {
   return [
     `[Playground](${document.location.toString()})`,
     createSummary(state.code, 'Code', 'ts', 30),
     createSummaryJson(state.eslintrc, 'rules', 'Eslint config'),
     createSummaryJson(state.tsconfig, 'compilerOptions', 'TypeScript config'),
-    genVersions(state),
+    genVersions(state.ts),
   ]
     .filter(Boolean)
     .join('\n\n');
 }
 
+/**
+ * Create a URLSearchParams string for the issue template
+ */
 export function createMarkdownParams(state: ConfigModel): string {
   const { rules } = parseESLintRC(state.eslintrc);
   const ruleKeys = Object.keys(rules);
@@ -66,7 +71,7 @@ export function createMarkdownParams(state: ConfigModel): string {
     'repro-code': state.code,
     'eslint-config': `module.exports = ${state.eslintrc ?? '{}'}`,
     'typescript-config': state.tsconfig ?? '{}',
-    versions: genVersions(state),
+    versions: genVersions(state.ts),
   };
 
   return new URLSearchParams(params).toString();
