+++ packages/website/src/components/lib/parseConfig.ts
@@ -1,17 +1,25 @@
-import type * as ts from 'typescript';
-
 import { isRecord } from '../ast/utils';
-import { parseJSONObject, toJson } from '../lib/json';
 import type { EslintRC, TSConfig } from '../types';
+import { ensureObject, parseJSONObject, toJson } from './json';
 
+/**
+ * Parse a .eslintrc string into an object.
+ * This function is pretty naive, as it only validates rules and extends.
+ */
 export function parseESLintRC(code?: string): EslintRC {
   if (code) {
     try {
       const parsed = parseJSONObject(code);
-      if ('rules' in parsed && isRecord(parsed.rules)) {
-        return parsed as EslintRC;
+      if ('rules' in parsed) {
+        parsed.rules = ensureObject(parsed.rules);
       }
-      return { ...parsed, rules: {} };
+      if ('extends' in parsed) {
+        parsed.extends =
+          Array.isArray(parsed.extends) || typeof parsed.extends === 'string'
+            ? parsed.extends
+            : [];
+      }
+      return parsed as EslintRC;
     } catch (e) {
       console.error(e);
     }
@@ -19,6 +27,10 @@ export function parseESLintRC(code?: string): EslintRC {
   return { rules: {} };
 }
 
+/**
+ * Parse a tsconfig.json string into an object.
+ * This is done by typescript compiler.
+ */
 export function parseTSConfig(code?: string): TSConfig {
   if (code) {
     try {
@@ -51,6 +63,9 @@ function constrainedScopeEval(obj: string): unknown {
   `)();
 }
 
+/**
+ * Evaluate a string that contains a module.exports assignment.
+ */
 export function tryParseEslintModule(value: string): string {
   try {
     if (moduleRegexp.test(value)) {
@@ -64,44 +79,3 @@ export function tryParseEslintModule(value: string): string {
   }
   return value;
 }
-
-export function getTypescriptOptions(): ts.OptionDeclarations[] {
-  const allowedCategories = [
-    'Command-line Options',
-    'Projects',
-    'Compiler Diagnostics',
-    'Editor Support',
-    'Output Formatting',
-    'Watch and Build Modes',
-    'Source Map Options',
-  ];
-
-  const filteredNames = [
-    'moduleResolution',
-    'moduleDetection',
-    'plugins',
-    'typeRoots',
-    'jsx',
-  ];
-
-  return window.ts.optionDeclarations.filter(
-    item =>
-      (item.type === 'boolean' ||
-        item.type === 'list' ||
-        item.type instanceof Map) &&
-      item.description &&
-      item.category &&
-      !allowedCategories.includes(item.category.message) &&
-      !filteredNames.includes(item.name),
-  );
-}
-
-export const defaultTsConfig = toJson({
-  compilerOptions: {
-    strictNullChecks: true,
-  },
-});
-
-export const defaultEslintConfig = toJson({
-  rules: {},
-});
