+++ packages/website/src/components/types.ts
@@ -6,7 +6,6 @@ export type CompilerFlags = Record<string, unknown>;
 export type SourceType = TSESLint.SourceType;
 
 export type RulesRecord = TSESLint.Linter.RulesRecord;
-export type RuleEntry = TSESLint.Linter.RuleEntry;
 
 export interface RuleDetails {
   name: string;
